requirements.txt
streamlit>=1.31
pandas>=2.0
numpy>=1.24
plotly>=5.18
scikit-learn>=1.3
xgboost>=2.0
google-api-python-client>=2.120
google-auth>=2.28
google-auth-httplib2>=0.2.0
